backend/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
astroid==3.3.8
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
Django==5.1.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Flask==3.0.2
isort==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.3.6
pygame==2.5.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
PyJWT==2.10.1
pylint==3.3.4
pywin32-ctypes==0.2.2
setuptools==69.1.1
sqlparse==0.5.1
tomlkit==0.13.2
tzdata==2024.1
Werkzeug==3.0.1
